refactor: add database setup checks in Makefile and rearrange codebase
