+++ frontend/src/managers/GameCustomization.ts
