+++ frontend/src/managers/TournamentManager.ts
